+++ src/eventsub/channel/channel_points_custom_reward_redemption/update.rs
@@ -29,8 +29,10 @@ impl EventSubscription for ChannelPointsCustomRewardRedemptionUpdateV1 {
 pub struct ChannelPointsCustomRewardRedemptionUpdateV1Payload {
     /// The requested broadcaster ID.
     pub broadcaster_user_id: types::UserId,
-    /// The requested broadcaster name.
-    pub broadcaster_user_name: types::UserName,
+    /// The requested broadcaster login.
+    pub broadcaster_user_login: types::UserName,
+    /// The requested broadcaster display name.
+    pub broadcaster_user_name: types::DisplayName,
     /// The redemption identifier.
     pub id: types::RewardId,
     /// RFC3339 timestamp of when the reward was redeemed.
@@ -41,10 +43,12 @@ pub struct ChannelPointsCustomRewardRedemptionUpdateV1Payload {
     pub status: RedemptionStatus,
     /// User ID of the user that redeemed the reward.
     pub user_id: types::UserId,
+    /// Login of the user that redeemed the reward.
+    pub user_login: types::UserName,
+    /// Display name of the user that redeemed the reward.
+    pub user_name: types::DisplayName,
     /// The user input provided. Empty string if not provided.
     pub user_input: String,
-    /// Display name of the user that redeemed the reward.
-    pub user_name: types::UserName,
 }
 
 #[test]
@@ -67,9 +71,11 @@ fn parse_payload() {
         "event": {
             "id": "1234",
             "broadcaster_user_id": "1337",
-            "broadcaster_user_name": "cool_user",
+            "broadcaster_user_login": "cool_user",
+            "broadcaster_user_name": "Cool_User",
             "user_id": "9001",
-            "user_name": "cooler_user",
+            "user_login": "cooler_user",
+            "user_name": "Cooler_User",
             "user_input": "pogchamp",
             "status": "fulfilled",
             "reward": {