+++ src/eventsub/channel/hypetrain/end.rs
@@ -28,8 +28,10 @@ impl EventSubscription for ChannelHypeTrainEndV1 {
 pub struct ChannelHypeTrainEndV1Payload {
     /// The requested broadcaster ID.
     pub broadcaster_user_id: types::UserId,
-    /// The requested broadcaster name.
-    pub broadcaster_user_name: types::UserName,
+    /// The requested broadcaster login.
+    pub broadcaster_user_login: types::UserName,
+    /// The requested broadcaster display name.
+    pub broadcaster_user_name: types::DisplayName,
     /// The timestamp at which the hype train cooldown ends so that the next hype train can start.
     pub cooldown_ends_at: types::Timestamp,
     /// The timestamp at which the hype train ended.
@@ -63,12 +65,13 @@ fn parse_payload() {
         },
         "event": {
             "broadcaster_user_id": "1337",
-            "broadcaster_user_name": "cool_user",
+            "broadcaster_user_login": "cool_user",
+            "broadcaster_user_name": "Cool_User",
             "level": 2,
             "total": 137,
             "top_contributions": [
-                { "user_id": "123", "user_name": "pogchamp", "type": "bits", "total": 50 },
-                { "user_id": "456", "user_name": "kappa", "type": "subscription", "total": 45 }
+                { "user_id": "123", "user_login": "pogchamp", "user_name": "PogChamp", "type": "bits", "total": 50 },
+                { "user_id": "456", "user_login": "kappa", "user_name": "Kappa", "type": "subscription", "total": 45 }
             ],
             "started_at": "2020-07-15T17:16:03.17106713Z",
             "ended_at": "2020-07-15T17:16:11.17106713Z",