+++ src/eventsub/channel/subscribe.rs
@@ -23,23 +23,27 @@ impl EventSubscription for ChannelSubscribeV1 {
     const VERSION: &'static str = "1";
 }
 
-/// [`channel.update`](ChannelSubscribeV1) response payload.
+/// [`channel.subscribe`](ChannelSubscribeV1) response payload.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[cfg_attr(not(feature = "allow_unknown_fields"), serde(deny_unknown_fields))]
 #[non_exhaustive]
 pub struct ChannelSubscribeV1Payload {
     /// The requested broadcaster ID.
     pub broadcaster_user_id: types::UserId,
-    /// The requested broadcaster name.
-    pub broadcaster_user_name: types::UserName,
+    /// The requested broadcaster login.
+    pub broadcaster_user_login: types::UserName,
+    /// The requested broadcaster display name.
+    pub broadcaster_user_name: types::DisplayName,
     /// Whether the subscription is a gift.
     pub is_gift: bool,
     /// The tier of the subscription. Valid values are 1000, 2000, and 3000.
     pub tier: types::SubscriptionTier,
     /// The user ID for the user who subscribed to the specified channel.
     pub user_id: types::UserId,
-    /// The user name for the user who subscribed to the specified channel.
-    pub user_name: types::UserName,
+    /// The user login for the user who subscribed to the specified channel.
+    pub user_login: types::UserName,
+    /// The user display name for the user who subscribed to the specified channel.
+    pub user_name: types::DisplayName,
 }
 
 #[test]
@@ -61,9 +65,11 @@ fn parse_payload() {
         },
         "event": {
             "user_id": "1234",
-            "user_name": "cool_user",
+            "user_login": "cool_user",
+            "user_name": "Cool_User",
             "broadcaster_user_id": "1337",
-            "broadcaster_user_name": "cooler_user",
+            "broadcaster_user_login": "cooler_user",
+            "broadcaster_user_name": "Cooler_User",
             "tier": "1000",
             "is_gift": false
         }